Finish project 6
Implement a "first pass", which just reads labels into the symbol table.
Use the symbol table during the second pass, to resolve symbols to their
corresponding memory addresses.

Note that allocating new variables for non-label symbols is done
*during* the second pass (code generation). This feels kinda clunky to
me -- but it would probably be overkill to split it into it's own
"middle pass".

testing: I ran it on the sample programs, and it produces the same
binary output as the reference implementation. So it seems to work!